Update svm.c
Got rid of "UPDATE" flags and all white space differences between this and 3.20's svm.cpp. There are still a lot more diffs between the two of which I will attempt to compare/change, but I just wanted this version out there which successfully makes. So just for clarification, this file is the result of comparing changes between the original libSVM's 3.1 and 3.20 and making the appropriate changes to Dr. Boult's version. There are still diffs between 3.20 and Dr. Boult's at this point, but I am unsure if changes need to be made still or not to get them more similar.